refactor(query): reorganize queries folder structure
