+++ apps/app/src/lib/query/_queries/transformationRuns.ts
@@ -1,6 +1,6 @@
 import { services } from '$lib/services';
 import type { Accessor } from '@tanstack/svelte-query';
-import { defineQuery } from '.';
+import { defineQuery } from '../_utils';
 
 // Define the query key as a constant array
 export const transformationRunKeys = {
