+++ apps/app/src/lib/query/_queries/vadRecorder.ts
@@ -4,9 +4,10 @@ import { settings } from '$lib/stores/settings.svelte';
 import { Ok, isOk } from '@epicenterhq/result';
 import type { WhisperingRecordingState } from '@repo/shared';
 import { nanoid } from 'nanoid/non-secure';
-import { defineMutation, defineQuery, queryClient } from '.';
+import { defineMutation, defineQuery } from '../_utils';
+import { queryClient } from '../index';
 import { recordings } from './recordings';
-import { maybeCopyAndPaste } from '../maybeCopyAndPaste';
+import { maybeCopyAndPaste } from '../../maybeCopyAndPaste';
 import { transcription } from './transcription';
 import { transformer } from './transformer';
